requirements .txt
Flask
torch
torchvision
openai
Pillow
